requirements.txt
flask
pygal
urllib3
Flask-WTF
HTMLParser
python-dateutil
virtualenv


